+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Next
 
+- Fix an issue where some historic transactions with removed bAssets would not show
+
 ## Version 1.16.0
 
 - Improve analytics page with 30 and 90 day views (including for APY), SAVE utilisation, bAsset vault balances,