+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Next
 
+## Version 1.16.1
+
 - Fix an issue where some historic transactions with removed bAssets would not show
 - Fix an issue where some of the basket stats were not displayed correctly